% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airbnb-class.R
\name{airbnb}
\alias{airbnb}
\title{Create an Airbnb object (S3 class)}
\usage{
airbnb(data)
}
\arguments{
\item{data}{A data frame containing Airbnb listings.}
}
\value{
An object of class \code{airbnb}.
}
\description{
Wraps an Airbnb dataset into an S3 object of class \code{"airbnb"}.
}
