% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_airbnb.R
\name{plot_airbnb}
\alias{plot_airbnb}
\title{Plot the Airbnb Price Distribution}
\usage{
plot_airbnb(x, ...)
}
\arguments{
\item{x}{An object of class \code{airbnb}.}

\item{...}{Additional arguments passed to \code{ggplot2}.}
}
\value{
A ggplot2 histogram.
}
\description{
Creates a histogram showing the distribution of listing prices.
}
