% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_by_roomtype.R
\name{plot_roomtype}
\alias{plot_roomtype}
\title{Plot Price by Room Type}
\usage{
plot_roomtype(object, ...)
}
\arguments{
\item{object}{An object of class \code{airbnb}.}

\item{...}{Additional arguments passed to ggplot2.}
}
\value{
A ggplot2 boxplot.
}
\description{
Creates a boxplot showing price differences by room type.
}
