% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_airbnb.R
\name{summary.airbnb}
\alias{summary.airbnb}
\title{Summary of an Airbnb Object}
\usage{
\method{summary}{airbnb}(object, ...)
}
\arguments{
\item{object}{An object of class \code{airbnb}.}

\item{...}{Additional arguments (unused).}
}
\value{
A list containing summary statistics such as average price,
median price, room type counts, and missing values.
}
\description{
Provides summary statistics for an Airbnb dataset.
}
